server/requirements.txt
attrs==23.1.0
blinker==1.6.2
click==8.1.3
exceptiongroup==1.1.2
flasgger==0.9.7.1
Flask==2.3.2
flask-pydantic-spec==0.4.5
gunicorn==20.1.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
MarkupSafe==2.1.3
mistune==3.0.1
numpy==1.25.0
packaging==23.1
pluggy==1.2.0
psutil==5.9.5
pydantic==1.10.10
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-mock==3.11.1
python-dotenv==1.0.0
PyYAML==6.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
